test: consolidate prefer-enum-initializers extras into a single file
